Implementation of low-priority Interest and ContentObject gradient pushing mechanism
When receiving Interest that cannot be satisfied or ContentObjects that
is not requested, nodes put the incoming packet to low-priority queue,
where waiting time depends on the distance from the original sender.

For waiting period randomization is used triangular distribution with
min 0, max 5ms (configurable through attributes), and mean that is
inversely proportional to distance (normalized to 300 meter distance,
which is also a configurable parameter).