apps: Solving a serious bug #33 in ndn::ConsumerWindow
In certain cases, the number of in-flight packets was maintained in a
wrong way, creating inconsistencies and incorrect decision on when to
send out a new Interests.  That is, the number of in-flight
packets could be non-zero, while there were no outstanding packets at all.

Thanks to Cheng Yi for the bug discovery