Use a single source of truth for InFlight calculation
ndn-consumer-window and ndn-consumer-pcon use a simple count for
updating the number of InFlight packets. This works fines as long as no
packet arrives after its timeout has fired. In this case `m_inFlight` is
decreased twice. Ultimately the counter (wrongly) reaches 0 and prevents
proper working of the window algorithm.

This fix does away with the manual calculation of the counter and replaces
it with the size of the `m_seqTimeouts` set.

Signed-off-by: Miguel Rodríguez Pérez <miguel@det.uvigo.gal>